test(core): add unit test for CEL expression ACK format
Added a unit test in `automation-manager.test.ts` to ensure that `send_packet` action correctly translates a string-based ACK (CEL expression) into the corresponding `StateSchema` with a `guard` property.
Also ran `pnpm format` as requested.
